+++ src/detection/displayserver/linux/displayserver_linux.c
@@ -32,14 +32,14 @@ static void parseDRM(FFDisplayServerResult* result)
             continue;
         }
 
-        FFResolutionResult* resolution = ffListAdd(&result->resolutions);
-        resolution->width = 0;
-        resolution->height = 0;
-        resolution->refreshRate = 0;
+        FFDisplayResult* display = ffListAdd(&result->displays);
+        display->width = 0;
+        display->height = 0;
+        display->refreshRate = 0;
 
-        int scanned = fscanf(modeFile, "%ux%u", &resolution->width, &resolution->height);
-        if(scanned < 2 || resolution->width == 0 || resolution->height == 0)
-            --result->resolutions.length;
+        int scanned = fscanf(modeFile, "%ux%u", &display->width, &display->height);
+        if(scanned < 2 || display->width == 0 || display->height == 0)
+            --result->displays.length;
 
         fclose(modeFile);
         ffStrbufSubstrBefore(&drmDir, drmDirLength);
@@ -57,31 +57,31 @@ void ffConnectDisplayServerImpl(FFDisplayServerResult* ds, const FFinstance* ins
     ffStrbufInit(&ds->deProcessName);
     ffStrbufInit(&ds->dePrettyName);
     ffStrbufInit(&ds->deVersion);
-    ffListInitA(&ds->resolutions, sizeof(FFResolutionResult), 4);
+    ffListInitA(&ds->displays, sizeof(FFDisplayResult), 4);
 
     //We try wayland as our prefered display server, as it supports the most features.
     //This method can't detect the name of our WM / DE
     ffdsConnectWayland(instance, ds);
 
     //Try the x11 libs, from most feature rich to least.
-    //We use the resolution list to detect if a connection is needed.
-    //They respect wmProtocolName, and only detect resolution if it is set.
+    //We use the display list to detect if a connection is needed.
+    //They respect wmProtocolName, and only detect display if it is set.
 
-    if(ds->resolutions.length == 0)
+    if(ds->displays.length == 0)
         ffdsConnectXcbRandr(instance, ds);
 
-    if(ds->resolutions.length == 0)
+    if(ds->displays.length == 0)
         ffdsConnectXrandr(instance, ds);
 
-    if(ds->resolutions.length == 0)
+    if(ds->displays.length == 0)
         ffdsConnectXcb(instance, ds);
 
-    if(ds->resolutions.length == 0)
+    if(ds->displays.length == 0)
         ffdsConnectXlib(instance, ds);
 
-    //This resolution detection method is display server independent.
+    //This display detection method is display server independent.
     //Use it if all connections failed
-    if(ds->resolutions.length == 0)
+    if(ds->displays.length == 0)
         parseDRM(ds);
 
     //This fills in missing information about WM / DE by using env vars and iterating processes