+++ src/detection/displayserver/linux/xlib.c
@@ -79,7 +79,7 @@ void ffdsConnectXlib(const FFinstance* instance, FFDisplayServerResult* result)
     for(int i = 0; i < ScreenCount(display); i++)
     {
         Screen* screen = ScreenOfDisplay(display, i);
-        ffdsAppendResolution(
+        ffdsAppendDisplay(
             result,
             (uint32_t) screen->width,
             (uint32_t) screen->height,
@@ -138,7 +138,7 @@ static bool xrandrHandleModeInfo(XrandrData* data, XRRModeInfo* modeInfo)
         modeInfo->dotClock / (modeInfo->hTotal * modeInfo->vTotal)
     ));
 
-    return ffdsAppendResolution(
+    return ffdsAppendDisplay(
         data->result,
         (uint32_t) modeInfo->width,
         (uint32_t) modeInfo->height,
@@ -158,7 +158,7 @@ static bool xrandrHandleMode(XrandrData* data, RRMode mode)
 
 static bool xrandrHandleCrtc(XrandrData* data, RRCrtc crtc)
 {
-    //We do the check here, because we want the best fallback resolution if this call failed
+    //We do the check here, because we want the best fallback display if this call failed
     if(data->screenResources == NULL)
         return false;
 
@@ -167,7 +167,7 @@ static bool xrandrHandleCrtc(XrandrData* data, RRCrtc crtc)
         return false;
 
     bool res = xrandrHandleMode(data, crtcInfo->mode);
-    res = res ? true : ffdsAppendResolution(
+    res = res ? true : ffdsAppendDisplay(
         data->result,
         (uint32_t) crtcInfo->width,
         (uint32_t) crtcInfo->height,
@@ -201,7 +201,7 @@ static bool xrandrHandleMonitor(XrandrData* data, XRRMonitorInfo* monitorInfo)
             foundOutput = true;
     }
 
-    return foundOutput ? true : ffdsAppendResolution(
+    return foundOutput ? true : ffdsAppendDisplay(
         data->result,
         (uint32_t) monitorInfo->width,
         (uint32_t) monitorInfo->height,
@@ -253,7 +253,7 @@ static void xrandrHandleScreen(XrandrData* data, Screen* screen)
         return;
 
     //Fallback to screen
-    ffdsAppendResolution(
+    ffdsAppendDisplay(
         data->result,
         (uint32_t) WidthOfScreen(screen),
         (uint32_t) HeightOfScreen(screen),